NSF validation: trailing nulls are required
Per http://wiki.nesdev.com/w/index.php/NSF#Header_Overview the name,
artist, and copyright fields all should be NULL-terminated.

This means, in ideal circumstances, that the last byte in each field
should be NULL.

One exception might be an NSF that does have a trailing NULL on the
string but chose to put other data after that, e.g. "abcdef<0x00>Hi"
(where "i" in "Hi" is the last byte in the field).  My opinion on
that is pretty simple: the NSF author is probably being a jerk.  :-)